[v] изменен формат Result и добавлены функции Unprocessed по аналогии с All

+++ grmqx/batch_handler/result.go
@@ -1,7 +1,7 @@
 package batch_handler
 
 const (
-	Unknown = iota
+	Unknown = int8(iota)
 	Ack
 	Retry
 	MoveToDlq
